+++ test/parser/test_groups.rb
@@ -20,8 +20,8 @@ class TestParserGroups < Test::Unit::TestCase
     assert_equal false, t.expressions[0].i?
     assert_equal false, t.expressions[0].x?
 
-    assert_equal true,  t.expressions[0].option_modifications[:m]
-    assert_equal nil,   t.expressions[0].option_modifications[:i]
+    assert_equal true,  t.expressions[0].option_changes[:m]
+    assert_equal nil,   t.expressions[0].option_changes[:i]
   end
 
   def test_parse_self_defeating_option_group
@@ -31,8 +31,8 @@ class TestParserGroups < Test::Unit::TestCase
     assert_equal false, t.expressions[0].i?
     assert_equal false, t.expressions[0].x?
 
-    assert_equal false, t.expressions[0].option_modifications[:m]
-    assert_equal nil,   t.expressions[0].option_modifications[:i]
+    assert_equal false, t.expressions[0].option_changes[:m]
+    assert_equal nil,   t.expressions[0].option_changes[:i]
   end
 
   def test_parse_nested_options_activate_one
@@ -46,9 +46,9 @@ class TestParserGroups < Test::Unit::TestCase
     assert_equal false, t.expressions[0].expressions[1].i?
     assert_equal true,  t.expressions[0].expressions[1].x?
 
-    assert_equal true,  t.expressions[0].expressions[1].option_modifications[:m]
-    assert_equal nil,   t.expressions[0].expressions[1].option_modifications[:i]
-    assert_equal nil,   t.expressions[0].expressions[1].option_modifications[:x]
+    assert_equal true,  t.expressions[0].expressions[1].option_changes[:m]
+    assert_equal nil,   t.expressions[0].expressions[1].option_changes[:i]
+    assert_equal nil,   t.expressions[0].expressions[1].option_changes[:x]
   end
 
   def test_parse_nested_options_deactivate_one
@@ -62,9 +62,9 @@ class TestParserGroups < Test::Unit::TestCase
     assert_equal false, t.expressions[0].expressions[1].i?
     assert_equal true,  t.expressions[0].expressions[1].x?
 
-    assert_equal false, t.expressions[0].expressions[1].option_modifications[:i]
-    assert_equal nil,   t.expressions[0].expressions[1].option_modifications[:m]
-    assert_equal nil,   t.expressions[0].expressions[1].option_modifications[:x]
+    assert_equal false, t.expressions[0].expressions[1].option_changes[:i]
+    assert_equal nil,   t.expressions[0].expressions[1].option_changes[:m]
+    assert_equal nil,   t.expressions[0].expressions[1].option_changes[:x]
   end
 
   def test_parse_nested_options_invert_all
@@ -78,9 +78,9 @@ class TestParserGroups < Test::Unit::TestCase
     assert_equal false, t.expressions[0].expressions[1].i?
     assert_equal false, t.expressions[0].expressions[1].x?
 
-    assert_equal true,  t.expressions[0].expressions[1].option_modifications[:m]
-    assert_equal false, t.expressions[0].expressions[1].option_modifications[:i]
-    assert_equal false, t.expressions[0].expressions[1].option_modifications[:x]
+    assert_equal true,  t.expressions[0].expressions[1].option_changes[:m]
+    assert_equal false, t.expressions[0].expressions[1].option_changes[:i]
+    assert_equal false, t.expressions[0].expressions[1].option_changes[:x]
   end
 
   def test_parse_nested_options_affect_literal_subexpressions
@@ -107,9 +107,9 @@ class TestParserGroups < Test::Unit::TestCase
     assert_equal true,  t.expressions[1].i?
     assert_equal false, t.expressions[1].x?
 
-    assert_equal true,  t.expressions[1].option_modifications[:i]
-    assert_equal false, t.expressions[1].option_modifications[:m]
-    assert_equal nil,   t.expressions[1].option_modifications[:x]
+    assert_equal true,  t.expressions[1].option_changes[:i]
+    assert_equal false, t.expressions[1].option_changes[:m]
+    assert_equal nil,   t.expressions[1].option_changes[:x]
   end
 
   def test_parse_option_switch_affects_following_expressions