+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Added
 
 - Added `Quantifier` methods `#greedy?`, `#possessive?`, `#reluctant?`/`#lazy?`
-- Added `Group::Options#option_modifications`
+- Added `Group::Options#option_changes`
   - shows the options enabled or disabled by the given options group
   - as with all other expressions, `#options` shows the overall active options
 - Added `Conditional#reference` and `Condition#reference`, indicating the determinative group